requirements.txt
streamlit
openai-whisper
faster-whisper
pydub
tqdm
requests

